increment, decrement cart and messaages flash

+++ src/Controller/CardController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller;
 
+use App\Repository\ProductRepository;
 use App\Service\Card\CardService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
@@ -23,19 +24,44 @@ class CardController extends AbstractController
     /**
      * @Route("/panier/add/{id}", name="card_add")
      */
-    public function add($id, CardService $cardService): Response
+    public function add($id, CardService $cardService, ProductRepository $productRepository): Response
     {
+        $product = $productRepository->find($id);
+        if(!$product) {
+            throw $this->createNotFoundException("Le product $id n'existe pas et ne peut pas être ajoute!");
+        }
         $cardService->add($id);
-
-       return $this->redirectToRoute("card_index");
+        $this->addFlash("success","Le produit a été bien ajouté");
+        return $this->redirectToRoute("card_index");
     }
 
     /**
      * @Route("/panier/remove/{id}", name="card_remove")
      */
-    public function remove($id, CardService $cardService)
+    public function remove($id, CardService $cardService, ProductRepository $productRepository)
     {
+        $product = $productRepository->find($id);
+        if(!$product) {
+            throw $this->createNotFoundException("Le product $id n'existe pas et ne peut pas être supprime!");
+        }
         $cardService->remove($id);
+        $this->addFlash("danger","Le produit a été bien supprime");
+        return $this->redirectToRoute("card_index");
+    }
+
+    /**
+     * @Route("/panier/decrement/{id}", name="card_decrement")
+     * @param $id
+     * @param CardService $cardService
+     */
+    public function decrement($id, CardService $cardService, ProductRepository $productRepository)
+    {
+        $product = $productRepository->find($id);
+        if(!$product) {
+            throw $this->createNotFoundException("Le product $id n'existe pas et ne peut pas etre décrémenté!");
+        }
+        $cardService->decrement($id);
+        $this->addFlash("danger","Le produit a été bien décrémenté");
         return $this->redirectToRoute("card_index");
     }
